Speed up browse searching.
Doing .last in AR runs a query which is "ORDER BY id DESC limit 1", but
this particular query in postgres reverses the entire index first and
then does an index scan. Since it's already using the sort index for the
WHERE, it can just use that for ordering, which will be in the proper
order.

This query runs in .095ms vs 2 seconds on the production DB.