Rescue and redirect from all BadRequest errors
Previously we were only rescuing and redirecting for a sub-set of
BadRequest errors. This meant that when there was a BadRequest that we
didn't anticipate, the user was getting an unhelpful error screen.
Better to always rescue, until we come up with some flavor of error that
we want to do something different for.