Remove mail_form gem
It is no longer needed.  This includes a refactor of the Report Biased Results
form, which is not in production and is behind a feature flipper.  That form
is not complete (for example, it does not have a honeypot field to catch spam
bots), I did not attempt to fix it, just kept the incomplete behavior the same.