[#4123] Avoid layout shift and quickly changing colors in the header
This commit moves the .pul_header container out of the Vue component
(which is rendered after the rest of the page), and into erb and
the application CSS, which will block rendering.  It also gives the
container an explicit height.

This prevents the situation where the page renders with light colored
content at first, then changes to very dark gray and shifts the layout
once the Vue component loads, which looks like a distracting flash.

Also, add a threshold for layout shift to our lighthouse check, to
hopefully guard against regressions.