Wrap our current index page partials in a subclass of Blacklight::DocumentComponent
Blacklight offers two ways to customize how your search results display:
* config.index.partials, where you can specify specific partials you want to show in the view directory
* config.index.document_component, where you can specify a subclass of Blacklight::DocumentComponent that
  displays the item

In Blacklight 8, the default behavior is to use Blacklight::DocumentComponent directly, which renders
all the relevant metadata for the result, and then display the partials you've configured in
config.index.partials.

The catch is that the partials we've configured in config.index.partials _also_ display the metadata,
so we end up displaying all the metadata twice.

This commit subclasses Blacklight::DocumentComponent to display only our partials.

Also, refactor _index_header_default into a subclass of Blacklight::DocumentTitleComponent, since
it made it easier to test.