[#4175] Extract Requestable::Holding class
Rather than passing around the holding as a hash throughout the application, this commit
begins an incremental refactor to pass around an object that can answer questions about
a holding from the holdings_1display. This will eventually mean that other classes won't
need to know the exact internal structure of the holding hash, they can just ask
Requestable::Holding for that information.

For now, though, this is just an incremental step, extracting a few methods from Requestable
into Requestable::Holding, and using the hash elsewhere in the application.

This reduces the reek count in Requestable by 2.