[#4372] Retrieving a temp location for the second time in Requests::Form should not return nil
Prior to this commit, Requests::Form#get_current_location would return the location data
the first time it was called for a particular temporary location, and then cache the hash in
memory.  However, the second time #get_current_location was called for a particular temp location,
rather than returning the cached value, it would return nil, which causes errors later in
the process.

This scenario happened when multiple items on a record are in the same temp location.

This commit extracts the caching logic into its own tiny class, and ensures that it returns the
location hash whether or not it had been cached previously.

Closes #4372