[#4344] Remove mm solr parameter when the user's query contains boolean OR
mm has a particular meaning when using the Boolean query parser, which is
used when we send JSON queries like this to solr:

{"query":{"bool":{"must":[{"edismax":{"query":"activism OR \"social justice\""}}]}}}

In these cases, according to Solr's documentation, mm specifies the
number of should clauses required for a document to be included in the
results.  In our configuration, the default mm is 6<90%, meaning that if
there are 6 or fewer clauses, all are required (and if there are 7 or more,
90% of the clauses are required).  This conflicts with and overrides the
OR that the user intended.

Therefore, if we notice that the user's query includes OR in all caps,
set the mm to 0 (meaning no minimum number of the clauses needs to be
included).

Many thanks to @maxkadel for troubleshooting this with me!

Closes #4344