[#4534] Make Bookmark All button compatible with Blacklight 8
Our jquery code attempts to trigger an event, but it appears that Blacklight 8 is not
listening to the events we send.  There are many changes in the JS between Blacklight 7
and Blacklight 8 (7 is in jquery while 8 is in vanilla js, 7 puts the event listener
on the specific checkbox/button while 8 puts the event listener on the entire document).

I'm not sure why the jquery event triggering doesn't work with the Blacklight 8 js, but
migrating from jquery to vanilla js seems to do the trick in both Blacklight 7 and 8.

Closes #4534