Remove a CSS customization that is no longer needed
We used to need a z-index to make sure that the library logo didn't cover the skip
links.  However, now that we use the Lux Header, we no longer need this!

We expecially won't need it in recent versions of Blacklight 8.  In those versions,
the skip link navigation pushes the rest of the content down, rather than appearing
on top of existing content.  Therefore, the skip links aren't even fighting for the
same part of the screen as the logo, so we definitely don't need a z-index.

In fact, this z-index customization looks very goofy on Blacklight 8.6.1.