[#4552] More reliable boolean operators on the advanced search form
This is not quite working, see failing test
spec/features/advanced_searching_spec.rb:156

Specifically, if the user enters a boolean NOT search in the advanced
search, the constraints on the search results page is currently not
including the word NOT, so the user doesn't get the feedback that it
is a negated term.

I think we could address this in our ClausePresenter, although it
might make that class quite a bit more complex.