[#4554] ClausePresenter: don't attempt to call #display_label on a non-existant field configuration
This fixes an error that appeared when a user specified a search field that
does not exist.  It is essentially the same as upstream fix https://github.com/projectblacklight/blacklight/pull/3442,
so if that is accepted and we start using a Blacklight release that includes
it, we can remove our customized version.