[#4595] Fix More in Series link on show page
Before we started using the JSON Query DSL, this link pointed to a
left-anchored search in a solr field called in_series.  The JSON
Query DSL was using a different field called series_title (thanks
@maxkadel for pointing that out), and it wasn't left-anchored.
This meant that users who clicked on that link would get a lot
of irrelevant records mixed in with titles from the series.

This commit brings back the old field and the old behavior into
the JSON Query DSL implementation.

Thanks to @minjiec for finding and reporting this issue!