[#4597] Add back all languages to facets on the advanced search form
Thanks @maxkadel for finding the underlying issue and writing failing tests!
This sends numismatics and advanced search forms through a different search
builder than other searches.  As a future step, it would be nice to move
some logic specific to those forms out of the general search builder and into
this one.