Simplify query for facet more link
When we use the "more" link on a facet, we send a
solr query to gather the facet values.  Prior to
this commit, we also sent some extra parameters
for other facets.  For example, we always sent
the facet.query params for the publication date and
cataloged date, even if we were looking at a
totally different facet, say, language.

Working locally with a branch for #4025 and catalog-staging,
the median speed for a suggestion query goes down from 2192 ms
to 1258 ms.

Helps with #4025, so that the feature is performant
enough to be useful