Add Requests::Submission#to_h method that is sidekiq-compatible
Sidekiq is very restrictive about which data structures we can pass in
with jobs, including to email jobs.

See https://github.com/sidekiq/sidekiq/wiki/Best-Practices#1-make-your-job-parameters-small-and-simple

Our Requests::Submission object was not built with sidekiq's restrictions
in mind, so this commit adds a method that can convert the object into
one that is compatible with Sidekiq.

Note that we do not yet use this method anywhere, this is just a preparatory
step.  The next steps will be to start incrementally teaching our mailer
views to use the to_h output rather than Requests::Submission directly.