Add a new bookmark button component
This bookmark button replaces the bookmark checkbox on the search
results page.

The first time a non-logged in user presses the button, they are
shown a dialog that encourages them to sign in.  The bookmark is
added, whether or not they choose to sign in.  We then record
the timestamp in localStorage, and do not show them the dialog
again as long as that localStorage is present.

Advances #4910
Advances #4927