Fix format icon centering on Firefox
This is a small follow-up to #5022.  Firefox and Chromium calculated the specifity
slightly differently, leading to Firefox getting display: block rather than
display: flex (and the icon was not centered as a result).

This commit removes the conflicting rule, so we can be confident that we'll have
display: flex in Firefox or Chromium.