Remove unused font css (#5053)
Just a small cleanup piece: the Droid Serif font is never used in the
catalog, so we don't need the @font-face rule for it.  At the end of
the day, this doesn't actually save any network calls, since modern
browsers are smart enough to not download unused fonts.  It still
offers a slight reduction in how much CSS the user needs to download
and that we need to maintain.