Keep holding group open when user returns from the request form
Before this commit, when a user opened a holding group on the show page, pressed the
Request button, then pressed the Return to Record button, the holding group was
no longer open.

This commit adds a param to the requests form that allows this Return button to
take the user back to the show page with the desired groups open.