Stop downloading franklin-gothic-urw font
Previously, we downloaded the font franklin-gothic-urw from Adobe,
and used it in a handful of lux components.

We are moving from franklin-gothic-urw to Libre Franklin, since
it is now recommended by the communications office and it is a
variable font.

Therefore, for the sake of consistency and saving the user some
bandwidth, we only use Libre Franklin.