[#5176] Don't display status badge on requests form for titles with 500+ items
Previously, rather than consulting the status on all 500+ items, we were defaulting
to the status message from the NullItem class, which is Unavailable.  In this case,
however, this was not a good default, since many/all items could be available, and
this could discourage users from requesting the item at all.

Thanks to @christinach for discussing this approach!

Closes #5176