Refactor: move Requestable creation out of Requests::Form
Prior to this commit, one of the Form class' many
responsibilities was to create many of the
Requestable instances needed by the Requests system.
It created four different kinds of Requestable, and
it was hard to trace the logic throughout the Form
class for each one of those four paths.

This commit takes a different approach: introducing
a factory class named RequestablesList, which
determines what type of Requestable to create,
then delegates the creation to one of four smaller
factories: AlmaItemsFactory, ScsbItemsFactory,
NoItemsFactory, and TooManyItemsFactory.