Retrieve patron data in its own Thread
For request forms with a single item, the performance bottleneck
is waiting for the patron information from Bibdata (typically
500-1500ms).

This commit places the Bibdata patron request into its own Thread,
so that we can do other work to construct the form while we
wait for the patron data.

In my local dev setup, the median amount of time taken to load
the Requests form is 300 ms faster on this branch than main.