Remove non-working focus management for modals
This code attempted to automatically focus the first input in a modal
when a modal was opened.  However, the event is never triggered in
practice, so this focus does not actually happen in any of the modal
forms I tried (SMS, Email, Ask a question, or Suggest a correction).

Let's remove this non-working code.

Helps with #5314 by removing jQuery code